api/services/container: define container service
After receiving feedback, it was found that we had to rethink the
dependence of a filesystem for containerd. This is a first pass at
redefining the request in this vein.

Specifically, we now take an opaque specification, along with a runtime
and mounts. The runtime will dispatch to execution backends, such as
runc. The specification is specific to the runtime, so the user still
gets access to the direct contents of the runc configuration.

Signed-off-by: Stephen J Day <stephen.day@docker.com>